refactor(model): separated visualizations of features and training
Separated the feature visualizations so that we can understand why we
use the features that we do and the training of the actual model.

Also serialized the scalars and the models that we are using so that
they can be called elsewhere in the project at ease with future API
plans.